passing item test
