+++ frontend/src/Components/ShoppingListItem/ItemDetails.js
@@ -4,11 +4,11 @@ import React from 'react';
 
 function ItemDetails({item}) {
     return (
-        <>
+        <React.Fragment>
             <td>{item.name}</td>
             <td>{item.quantity}</td>
             <td>{item.unit}</td>
-        </>
+        </React.Fragment>
     )
 }
